Gesture eventing fixed
The gesture eventing was not setup correctly such that touchables could still get events inside of the swipeable view.

Using the view’s Capture functions instead of the PanResponder dosn’t give as much info so those props were removed and moved to the PanResponders create method.  onStartShouldSetPanResponderCapture did not seem necessary.  onMoveShouldSetPanResponder seemed to never be called and its logic was moved to onMoveShouldSetPanResponderCapture that IS called when a touchable is pressed/gestured inside the swipable view such that sensitivity can be taken into account.