Remove Promises from open functions
Attempting to fix state problem with removing Promises from openRight and openLeft and instead using the setState callback to call the tweenContent and set the final state.

Currently the issue is that if an open button is used to open the swipe menu, the swipe state is now tied together with other swipe menus on swipe.