fix: export right interface with right props
The package is exporting an interface SwipeoutProperties from 'index.d.ts',
this interface has a prop named 'rowId' but this prop is not supported in the
package. The prop that is supported in package is 'rowID'. Because of this user
was getting wrong intellisense in the editor and even after updating row id in props
it wasn't updated